+++ internal/app/process/country_test.go
@@ -32,8 +32,8 @@ func TestProcess(t *testing.T) {
 
 		requester.On("Countries").Return(ch)
 
-		repo.On("GetById", 180).Return(&app.Country{}, errors.New("not Found"))
-		repo.On("GetById", 5).Return(&app.Country{}, errors.New("not Found"))
+		repo.On("GetById", int64(180)).Return(&app.Country{}, errors.New("not Found"))
+		repo.On("GetById", int64(5)).Return(&app.Country{}, errors.New("not Found"))
 		repo.On("Insert", eng).Return(nil)
 		repo.On("Insert", ger).Return(nil)
 
@@ -66,8 +66,8 @@ func TestProcess(t *testing.T) {
 
 		requester.On("Countries").Return(ch)
 
-		repo.On("GetById", 180).Return(eng, nil)
-		repo.On("GetById", 5).Return(ger, nil)
+		repo.On("GetById", int64(180)).Return(eng, nil)
+		repo.On("GetById", int64(5)).Return(ger, nil)
 		repo.On("Update", &eng).Return(nil)
 		repo.On("Update", &ger).Return(nil)
 
@@ -100,8 +100,8 @@ func TestProcess(t *testing.T) {
 
 		requester.On("Countries").Return(ch)
 
-		repo.On("GetById", 180).Return(&app.Country{}, errors.New("not Found"))
-		repo.On("GetById", 5).Return(&app.Country{}, errors.New("not Found"))
+		repo.On("GetById", int64(180)).Return(&app.Country{}, errors.New("not Found"))
+		repo.On("GetById", int64(5)).Return(&app.Country{}, errors.New("not Found"))
 		repo.On("Insert", eng).Return(errors.New("error occurred"))
 		repo.On("Insert", ger).Return(nil)
 
@@ -135,8 +135,8 @@ func TestProcess(t *testing.T) {
 
 		requester.On("Countries").Return(ch)
 
-		repo.On("GetById", 180).Return(eng, nil)
-		repo.On("GetById", 5).Return(ger, nil)
+		repo.On("GetById", int64(180)).Return(eng, nil)
+		repo.On("GetById", int64(5)).Return(ger, nil)
 		repo.On("Update", &eng).Return(errors.New("error occurred"))
 		repo.On("Update", &ger).Return(nil)
 
@@ -151,7 +151,7 @@ func TestProcess(t *testing.T) {
 	})
 }
 
-func newCountry(id int, name string) *app.Country {
+func newCountry(id int64, name string) *app.Country {
 	c := app.Country{
 		ID:        id,
 		Name:      name,
